flasher: Tidy up by moving the ihex handler in with the firmware reading logic
